+++ src/main/java/dev/vorstu/services/TeacherService.java
@@ -4,6 +4,7 @@ import dev.vorstu.dto.GroupNameDTO;
 import dev.vorstu.dto.TeacherDTO;
 import dev.vorstu.entities.GroupEntity;
 import dev.vorstu.entities.TeacherEntity;
+import dev.vorstu.mappers.GroupMapper;
 import dev.vorstu.mappers.TeacherMapper;
 import dev.vorstu.repositories.GroupRepository;
 import dev.vorstu.repositories.TeacherRepository;
@@ -20,36 +21,42 @@ public class TeacherService {
     private final TeacherRepository teacherRepository;
     private final TeacherMapper teacherMapper;
     private final GroupRepository groupRepository;
+    private final GroupMapper groupMapper;
 
+    public TeacherDTO addTeacher(TeacherDTO teacherDTO){
 
-    // TODO допроверить логику
-    public void addTeacher(TeacherDTO teacherDTO){
+        if(teacherRepository.existsById(teacherDTO.getId())){
+            throw new RuntimeException("Учителя с id - " + teacherDTO.getId() + " уже существует");
+        }
+
         TeacherEntity teacherEntity = teacherMapper.toTeacherEntity(teacherDTO);
         if(teacherDTO.getGroupsOfStudents() != null){
             teacherEntity.setGroupsOfStudents(toListGroupEntities(teacherDTO.getGroupsOfStudents()));
         }
-        teacherRepository.save(teacherEntity);
+        return teacherMapper.toTeacherDTO(teacherRepository.save(teacherEntity));
     }
 
-    // TODO по-хорошему проверка на успешное удаление
     public Long deleteTeacher(Long id){
+
+        if(!teacherRepository.existsById(id)){
+            throw new RuntimeException("Учителя с id - " + id + " не существует");
+        }
+
         teacherRepository.deleteById(id);
         return id;
     }
 
-    // TODO через маппер должно работать
     public TeacherDTO updateTeacher(TeacherDTO teacherDTO){
-        TeacherEntity teacherEntity = teacherRepository.findById(teacherDTO.getId()).orElseThrow();
-        teacherEntity.setName(teacherDTO.getName());
-        teacherEntity.setSurname(teacherDTO.getSurname());
-        teacherEntity.setPatronymic(teacherDTO.getPatronymic());
+
+        if(!teacherRepository.existsById(teacherDTO.getId())){
+            throw new RuntimeException("Учителя с id - " + teacherDTO.getId() + " не существует");
+        }
+
+        TeacherEntity teacherEntity = teacherMapper.toTeacherEntity(teacherDTO);
         teacherEntity.setGroupsOfStudents(toListGroupEntities(teacherDTO.getGroupsOfStudents()));
-        teacherRepository.save(teacherEntity);
-        return teacherDTO;
+        return teacherMapper.toTeacherDTO(teacherRepository.save(teacherEntity));
     }
 
-
-    // TODO по логике должен вроде быть в маппере, но нужен репозиторий, поэтому здесь
     public List<GroupEntity> toListGroupEntities(List<GroupNameDTO> groupNameDTOList) {
 
         List<Long> ids = new ArrayList<>();
@@ -58,6 +65,23 @@ public class TeacherService {
             ids.add(group.getId());
         }
         return groupRepository.getGroupsByIds(ids);
+    }
+
+    public TeacherDTO getTeacherById(Long id){
+
+        if(id == null){
+            throw new RuntimeException("Id не может быть равно нулю");
+        }
+        TeacherEntity teacherEntity = teacherRepository.findById(id).orElseThrow();
+        return teacherMapper.toTeacherDTO(teacherEntity);
+    }
+
+    public List<GroupNameDTO> getTeacherGroups(Long teacherId){
+        if (teacherId == null){
+            throw new RuntimeException("Id не может быть равно нулю");
+        }
+        List<GroupEntity> groups = this.teacherRepository.getTeacherGroups(teacherId).orElseThrow();
+        return this.groupMapper.toListGroupNameDTOs(groups);
     }
 
 }
